Fix the Oracle provider, broken since v4.5.10
The GetLastInsertIdSqlSuffix<T>() interface method introduced in v4.5.10 broke the Oracle provider.

The method was changed from a side-effect (returning a SQL phrase that could be appended to a command performing an INSERT) to a method evaluating the INSERT operation itself.

This provides enough context for the Oracle provider to perform its retrieval of the inserted ID.